Support capabilities from CRI with LCOW
This change adds support to LCOW for the `capabilities` field on CRI's
`LinuxContainerSecurityContext`. The code added in container_create_windows.go
to support this is identical to what exists in container_create_unix.go.
